% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{replace_na}
\alias{replace_na}
\title{Replace NA values with "NA" value}
\usage{
replace_na(x)
}
\arguments{
\item{x}{A vector or value to check for NA.}
}
\value{
"NA" if input was NA.
}
\description{
This utility function catches NA values inside structural inputs and
replaces them with an "NA".
}
\examples{
replace_na(NA)
replace_na("Valid String")
}
